Refactor: DTO 구조 변경
