+++ src/main/java/com/sprint/mission/discodeit/controller/dto/user/UserIdResponse.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.service.dto.user;
+package com.sprint.mission.discodeit.controller.dto.user;
 
 import java.util.UUID;
