+++ src/main/java/com/sprint/mission/discodeit/controller/dto/user/UserStatusUpdateDataRequest.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.controller.dto;
+package com.sprint.mission.discodeit.controller.dto.user;
 
 public record UserStatusUpdateDataRequest(
         String status
