+++ src/main/java/com/sprint/mission/discodeit/controller/dto/user/UserStatusUpdateResponse.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.controller.dto;
+package com.sprint.mission.discodeit.controller.dto.user;
 
 import com.sprint.mission.discodeit.entity.UserStatusType;
